requirements.txt
requests==2.32.3
beautifulsoup4==4.12.3
pdfplumber==0.11.4
langchain[google-genai]  # This installs latest langchain + langchain-google-genai + google-genai SDK
qdrant-client==1.16.2
pyyaml==6.0.2
python-dotenv==1.0.1